+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### Visual changes
 * Added setting for empty label text
 * Removed the exclusion of empty data from the legend, allowing users to leverage the native "Show items with no data" functionality.
+### Fixes
+* Fix sorting for the tasks
+
 
 ## 3.4.2
 ### Translations